requirements.txt
crewai[litellm]>=1.10.0
python-dotenv>=1.0.0
requests>=2.31.0
duckduckgo-search>=6.0.0
